add connector for spark3 & extract common code
