+++ nebula-spark-common/pom.xml
@@ -22,7 +22,7 @@
 
         <scala.binary.version>2.11</scala.binary.version>
         <spark.version>2.4.4</spark.version>
-        <scala.version>2.11.12</scala.version>
+        <scala.version>2.12.10</scala.version>
     </properties>
 
     <dependencies>
