+++ nebula-spark-connector_3.0/src/main/scala/com/vesoft/nebula/connector/NebulaDataSource.scala
@@ -50,7 +50,7 @@ class NebulaDataSource extends TableProvider with DataSourceRegister {
     if (schema == null) {
       nebulaOptions = getNebulaOptions(caseInsensitiveStringMap)
       if (nebulaOptions.operaType == OperaType.READ) {
-        schema = getSchema(nebulaOptions)
+        schema = NebulaUtils.getSchema(nebulaOptions)
       } else {
         schema = new StructType()
       }
@@ -62,72 +62,6 @@ class NebulaDataSource extends TableProvider with DataSourceRegister {
                         transforms: Array[Transform],
                         map: util.Map[String, String]): Table = {
     new NebulaTable(tableSchema, nebulaOptions)
-  }
-
-  /**
-    * return the dataset's schema. Schema includes configured cols in returnCols or includes all properties in nebula.
-    */
-  private def getSchema(nebulaOptions: NebulaOptions): StructType = {
-    val returnCols                      = nebulaOptions.getReturnCols
-    val noColumn                        = nebulaOptions.noColumn
-    val fields: ListBuffer[StructField] = new ListBuffer[StructField]
-    val metaProvider = new MetaProvider(
-      nebulaOptions.getMetaAddress,
-      nebulaOptions.timeout,
-      nebulaOptions.connectionRetry,
-      nebulaOptions.executionRetry,
-      nebulaOptions.enableMetaSSL,
-      nebulaOptions.sslSignType,
-      nebulaOptions.caSignParam,
-      nebulaOptions.selfSignParam
-    )
-
-    import scala.collection.JavaConverters._
-    var schemaCols: Seq[ColumnDef] = Seq()
-    val isVertex                   = DataTypeEnum.VERTEX.toString.equalsIgnoreCase(nebulaOptions.dataType)
-
-    // construct vertex or edge default prop
-    if (isVertex) {
-      fields.append(DataTypes.createStructField("_vertexId", DataTypes.StringType, false))
-    } else {
-      fields.append(DataTypes.createStructField("_srcId", DataTypes.StringType, false))
-      fields.append(DataTypes.createStructField("_dstId", DataTypes.StringType, false))
-      fields.append(DataTypes.createStructField("_rank", DataTypes.LongType, false))
-    }
-
-    var dataSchema: StructType = null
-    // read no column
-    if (noColumn) {
-      dataSchema = new StructType(fields.toArray)
-      return dataSchema
-    }
-    // get tag schema or edge schema
-    val schema = if (isVertex) {
-      metaProvider.getTag(nebulaOptions.spaceName, nebulaOptions.label)
-    } else {
-      metaProvider.getEdge(nebulaOptions.spaceName, nebulaOptions.label)
-    }
-
-    schemaCols = schema.columns.asScala
-
-    // read all columns
-    if (returnCols.isEmpty) {
-      schemaCols.foreach(columnDef => {
-        LOG.info(s"prop name ${new String(columnDef.getName)}, type ${columnDef.getType.getType} ")
-        fields.append(
-          DataTypes.createStructField(new String(columnDef.getName),
-                                      NebulaUtils.convertDataType(columnDef.getType),
-                                      true))
-      })
-    } else {
-      for (col: String <- returnCols) {
-        fields.append(
-          DataTypes
-            .createStructField(col, NebulaUtils.getColDataType(schemaCols.toList, col), true))
-      }
-    }
-    dataSchema = new StructType(fields.toArray)
-    dataSchema
   }
 
   /**
